movie-ticket-system/movie-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose
passlib